+++ examples/lox/src/language/lox-type-checking.ts
@@ -5,7 +5,7 @@
 ******************************************************************************/
 
 import { AstNode, AstUtils, LangiumSharedCoreServices, Module, assertUnreachable } from 'langium';
-import { CreateFieldDetails, CreateMethodDetails, CreateParameterDetails, FunctionType, InferOperatorWithMultipleOperands, InferOperatorWithSingleOperand, InferenceRuleNotApplicable, NO_PARAMETER_NAME, TypeInitializer, TypirServices, UniqueClassValidation, UniqueFunctionValidation, UniqueMethodValidation, ValidationMessageDetails, ValidationProblem, createNoSuperClassCyclesValidation } from 'typir';
+import { CreateFieldDetails, CreateMethodDetails, CreateParameterDetails, FunctionType, InferOperatorWithMultipleOperands, InferOperatorWithSingleOperand, InferenceRuleNotApplicable, NO_PARAMETER_NAME, TypeInitializer, TypirServices, UniqueClassValidation, UniqueFunctionValidation, UniqueMethodValidation, ValidationMessageDetails, ValidationProblemAcceptor, createNoSuperClassCyclesValidation } from 'typir';
 import { AbstractLangiumTypeCreator, LangiumLanguageService, LangiumServicesForTypirBinding, PartialTypirLangiumServices } from 'typir-langium';
 import { BinaryExpression, BooleanLiteral, Class, ForStatement, FunctionDeclaration, IfStatement, LoxAstType, MemberCall, MethodMember, NilLiteral, NumberLiteral, PrintStatement, ReturnStatement, StringLiteral, TypeReference, UnaryExpression, WhileStatement, isClass, isFieldMember, isFunctionDeclaration, isMemberCall, isMethodMember, isParameter, isVariableDeclaration, reflection } from './generated/ast.js';
 
@@ -87,7 +87,7 @@ export class LoxTypeCreator extends AbstractLangiumTypeCreator {
                 .inferenceRule({
                     ...binaryInferenceRule,
                     // show a warning to the user, if something like "3 == false" is compared, since different types already indicate, that the IF condition will be evaluated to false
-                    validation: (node, _operatorName, _operatorType, typir) => typir.validation.Constraints.ensureNodeIsEquals(node.left, node.right, (actual, expected) => <ValidationMessageDetails>{
+                    validation: (node, _operatorName, _operatorType, accept, typir) => typir.validation.Constraints.ensureNodeIsEquals(node.left, node.right, accept, (actual, expected) => <ValidationMessageDetails>{
                         message: `This comparison will always return '${node.operator === '==' ? 'false' : 'true'}' as '${node.left.$cstNode?.text}' and '${node.right.$cstNode?.text}' have the different types '${actual.name}' and '${expected.name}'.`,
                         languageNode: node, // inside the BinaryExpression ...
                         languageProperty: 'operator', // ... mark the '==' or '!=' token, i.e. the 'operator' property
@@ -102,7 +102,7 @@ export class LoxTypeCreator extends AbstractLangiumTypeCreator {
             .inferenceRule({
                 ...binaryInferenceRule,
                 // this validation will be checked for each call of this operator!
-                validation: (node, _opName, _opType, typir) => typir.validation.Constraints.ensureNodeIsAssignable(node.right, node.left, (actual, expected) => <ValidationMessageDetails>{
+                validation: (node, _opName, _opType, accept, typir) => typir.validation.Constraints.ensureNodeIsAssignable(node.right, node.left, accept, (actual, expected) => <ValidationMessageDetails>{
                     message: `The expression '${node.right.$cstNode?.text}' of type '${actual.name}' is not assignable to '${node.left.$cstNode?.text}' with type '${expected.name}'`,
                     languageProperty: 'value' })})
             .finish();
@@ -155,29 +155,26 @@ export class LoxTypeCreator extends AbstractLangiumTypeCreator {
         this.typir.validation.Collector.addValidationRulesForAstNodes<LoxAstType>({
             ForStatement: validateCondition,
             IfStatement: validateCondition,
-            ReturnStatement: (node, typir) => {
+            ReturnStatement: (node, accept, typir) => {
                 const callableDeclaration: FunctionDeclaration | MethodMember | undefined = AstUtils.getContainerOfType(node, node => isFunctionDeclaration(node) || isMethodMember(node));
                 if (callableDeclaration && callableDeclaration.returnType.primitive && callableDeclaration.returnType.primitive !== 'void' && node.value) {
                     // the return value must fit to the return type of the function / method
-                    return typir.validation.Constraints.ensureNodeIsAssignable(node.value, callableDeclaration.returnType, (actual, expected) => <ValidationMessageDetails<AstNode>>{
+                    typir.validation.Constraints.ensureNodeIsAssignable(node.value, callableDeclaration.returnType, accept, (actual, expected) => <ValidationMessageDetails<AstNode>>{
                         message: `The expression '${node.value!.$cstNode?.text}' of type '${actual.name}' is not usable as return value for the function '${callableDeclaration.name}' with return type '${expected.name}'.`,
                         languageProperty: 'value' });
                 }
-                return []; // TODO Review: this line is a bit ugly, is there a way to get rid of it?
             },
-            VariableDeclaration: (node, typir) => {
-                return [
-                    ...typir.validation.Constraints.ensureNodeHasNotType(node, typeVoid,
-                        () => <ValidationMessageDetails<AstNode>>{ message: "Variable can't be declared with a type 'void'.", languageProperty: 'type' }),
-                    ...typir.validation.Constraints.ensureNodeIsAssignable(node.value, node, (actual, expected) => <ValidationMessageDetails<AstNode>>{
-                        message: `The expression '${node.value?.$cstNode?.text}' of type '${actual.name}' is not assignable to '${node.name}' with type '${expected.name}'`,
-                        languageProperty: 'value' }),
-                ];
+            VariableDeclaration: (node, accept, typir) => {
+                typir.validation.Constraints.ensureNodeHasNotType(node, typeVoid, accept,
+                    () => <ValidationMessageDetails<AstNode>>{ message: "Variable can't be declared with a type 'void'.", languageProperty: 'type' });
+                typir.validation.Constraints.ensureNodeIsAssignable(node.value, node, accept, (actual, expected) => <ValidationMessageDetails<AstNode>>{
+                    message: `The expression '${node.value?.$cstNode?.text}' of type '${actual.name}' is not assignable to '${node.name}' with type '${expected.name}'`,
+                    languageProperty: 'value' });
             },
             WhileStatement: validateCondition,
         });
-        function validateCondition(node: IfStatement | WhileStatement | ForStatement, typir: TypirServices<AstNode>): Array<ValidationProblem<AstNode>> {
-            return typir.validation.Constraints.ensureNodeIsAssignable(node.condition, typeBool,
+        function validateCondition(node: IfStatement | WhileStatement | ForStatement, accept: ValidationProblemAcceptor<AstNode>, typir: TypirServices<AstNode>): void {
+            typir.validation.Constraints.ensureNodeIsAssignable(node.condition, typeBool, accept,
                 () => <ValidationMessageDetails<AstNode>>{ message: "Conditions need to be evaluated to 'boolean'.", languageProperty: 'condition' });
         }
         // TODO Review: Should all checks be moved into functions? Should all these validations moved into another file?