+++ packages/typir/test/kinds/function/operator-overloaded.test.ts
@@ -57,12 +57,11 @@ describe('Multiple best matches for overloaded operators', () => {
         });
 
         // register a type-related validation
-        typir.validation.Collector.addValidationRule(node => {
+        typir.validation.Collector.addValidationRule((node, accept) => {
             if (node instanceof AssignmentStatement) {
-                return typir.validation.Constraints.ensureNodeIsAssignable(node.right, node.left, (actual, expected) => <ValidationMessageDetails<TestLanguageNode>>{ message:
+                typir.validation.Constraints.ensureNodeIsAssignable(node.right, node.left, accept, (actual, expected) => <ValidationMessageDetails<TestLanguageNode>>{ message:
                     `The type '${actual.name}' is not assignable to the type '${expected.name}'.` });
             }
-            return [];
         });
     });